[FLINK-17308] Extract ExecutionGraphCache interface and rename impl into DefaultExecutionGraphCache
This commit extracts the ExecutionGraphCache interface from the implementation and renames
the latter into DefaultExecutionGraphCache. Moreover, it introduces the NoOpExecutionGraphCache
implementation which is used for the DocumentingDispatcherRestEndpoint.
